+++ bin/setup.sh
@@ -1,0 +1,8 @@
+# simple stap setup project
+composer install
+cp .env.example .env
+npm install
+npm run dev
+
+# builit cli
+php CLI --help