+++ bin/update.sh
@@ -1,0 +1,6 @@
+composer install
+npm install
+npm run dev
+
+# builit cli
+php CLI --help