+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.1] - 2021-02-05
+### Add
+- Add setup.sh, short hand to setup project
+
+### Changed
+- Refactory MyQuery
+
 ## [0.1.0] - 2021-01-28
 ### Added
 - Project init